backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
beanie==1.27.0
motor==3.5.1
pymongo==4.8.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
minio==7.2.8
pillow==10.4.0
bleach==6.1.0
pytz==2025.1
slowapi==0.1.9
pydantic[email]==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.3
httpx==0.27.0
